+++ src/objects/EntityFactory.cpp
@@ -7,7 +7,7 @@ EntityFactory::EntityFactory() {
 EntityFactory::~EntityFactory() {
 }
 
-double EntityFactory::getSpecSize(SpacingType spacing) {
+float EntityFactory::getSpecSize(SpacingType spacing) {
 	switch (spacing) {
 	case CONSTANT:
 	case RANDOM:
